+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Pre-release changes - please put everything in the appropriate category below
 
 ### Enhancements
+- no-ticket - Disabled case study tagging search ability on Admin UI
 - no-ticket - update placeholders
 - GP2-2773 - Update sign-in & sign-up with new logo
 - GP2-2783 - Product search re-label rework